+++ bk2ch16p687AVQueuePlayer/ch29p952AVQueuePlayer/Player.m
@@ -50,7 +50,7 @@
 }
 
 -(void)audioPlayerEndInterruption:(AVAudioPlayer *)player withOptions:(NSUInteger)opts {
-    NSLog(@"audio player ended interruption options: %i", opts);
+    NSLog(@"audio player ended interruption options: %lu", (unsigned long)opts);
 }
